+++ src/XRP/model/xrp-transaction.ts
@@ -22,7 +22,10 @@ export default class XRPTransaction {
       return undefined
     }
 
-    const account = transaction.getAccount()?.getValue()?.getAddress()
+    const account = transaction
+      .getAccount()
+      ?.getValue()
+      ?.getAddress()
 
     const fee = transaction.getFee()?.getDrops()
 
@@ -91,6 +94,22 @@ export default class XRPTransaction {
         ? rippleTransactionDate + 946684800
         : undefined
 
+    const deliveredAmountProto = getTransactionResponse
+      .getMeta()
+      ?.getDeliveredAmount()
+
+    const deliveredAmountXRP = deliveredAmountProto
+      ?.getValue()
+      ?.getXrpAmount()
+      ?.getDrops()
+
+    const deliveredAmountIssuedCurrency = deliveredAmountProto
+      ?.getValue()
+      ?.getIssuedCurrencyAmount()
+      ?.getValue()
+
+    const deliveredAmount = deliveredAmountXRP ?? deliveredAmountIssuedCurrency
+
     return new XRPTransaction(
       transactionHash,
       account,
@@ -107,6 +126,7 @@ export default class XRPTransaction {
       type,
       paymentFields,
       timestamp,
+      deliveredAmount,
     )
   }
 
@@ -126,5 +146,6 @@ export default class XRPTransaction {
     readonly type?: XRPTransactionType,
     readonly paymentFields?: XRPPayment,
     readonly timestamp?: number,
+    readonly deliveredAmount?: string,
   ) {}
 }