+++ test/PayID/pay-id-admin-client-test.ts
@@ -7,8 +7,8 @@ const xrpAddress = 'X7cBcY4bdTTzk3LHmrKAK6GyrirkXfLHGFxzke5zTmYMfw4'
 const serviceURL = 'https://xpring.money'
 const authorizationToken = 'fake-bearer-token'
 
-describe('Pay ID Admin Client', function (): void {
-  it('addOrUpdateXRPAddressMapping - throws unimplemented', function (done): void {
+describe('Pay ID Admin Client', function(): void {
+  it('addOrUpdateXRPAddressMapping - throws unimplemented', function(done): void {
     // GIVEN a PayIDAdminClient.
     const payIDAdminClient = new PayIDAdminClient(
       authorizationToken,
@@ -27,7 +27,7 @@ describe('Pay ID Admin Client', function (): void {
       })
   })
 
-  it('removeXRPAddressMapping - throws unimplemented', function (done): void {
+  it('removeXRPAddressMapping - throws unimplemented', function(done): void {
     // GIVEN a PayIDAdminClient.
     const payIDAdminClient = new PayIDAdminClient(
       authorizationToken,