+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+# Added
+- `XRPTransaction` contains additional field `deliveredAmount` to represent the actual delivered amount of a transaction.
 
 ## 4.2.0 - April 7, 2020